# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Emergency Dispatch System Backend API
  version: 1.0.0
  description: |
    Comprehensive OpenAPI specification for all backend services in the
    Emergency Dispatch System.

    Services covered:
    - Auth Service
    - Incident Service
    - Dispatch Service
    - Analytics Service
    - Realtime Gateway (WebSocket handshake endpoint)
servers:
  - url: http://localhost:8081
    description: Auth service (local)
  - url: http://localhost:8082
    description: Incident service (local)
  - url: http://localhost:8083
    description: Dispatch service (local)
  - url: http://localhost:8084
    description: Analytics service (local)
  - url: http://localhost:8085
    description: Realtime gateway (local)
...
security:
  - bearerAuth: []
paths:
...
  /auth/refresh-token:
    post:
      tags: [Auth]
      summary: Refresh token pair
      security: []
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Refreshed tokens issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/profile:
    get:
      tags: [Auth]
      summary: Get current user profile
      operationId: getProfile
      responses:
        '200':
          description: Current profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Auth]
      summary: Update current user profile
      operationId: updateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/users:
    get:
      tags: [Auth]
      summary: List users (system_admin)
      operationId: listUsers
      responses:
        '200':
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /auth/users/{id}:
    parameters:
      - $ref: '#/components/parameters/UUIDPathId'
    put:
      tags: [Auth]
      summary: Update user (system_admin)
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Auth]
      summary: Delete user (system_admin)
      operationId: deleteUser
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UUIDPathId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    RefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string
      required: [refresh_token]
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          format: int64
      required: [access_token, refresh_token, expires_in]
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
        station_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        password:
          type: string
          minLength: 8
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum:
            - system_admin
            - hospital_admin
            - police_admin
            - fire_admin
            - ambulance_driver
            - police_driver
            - fire_driver
        station_id:
          type: string
          format: uuid
          nullable: true
...